requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0
plotly>=5.15.0
yfinance>=0.2.18
fredapi>=0.5.1
python-dotenv>=1.0.0
scipy>=1.10.0
